* ffmpeg_track/release-1.0: (26 commits)
  update for 1.0.1
  qdm2: check array index before use, fix out of array accesses
  huffyuvdec: check width more completely, avoid out of array accesses
  roqvideodec: check dimensions validity
  h264: correct ref count check and limit, fix out of array accesses.
  eamad: fix out of array accesses
  pgssubdec: check RLE size before copying. Fix out of array accesses
  wavpack: fix out of array access
  mjpegdec: check SE.
  alac: fix integer overflow leading to subsequent out of array accesses.
  alac: fix nb_samples < order case
  iff/ilbm: check remaining buffer size.
  tiffdec: check count in metadata reading.
  swr: use 64bit index/len for x86_64 mix
  mpegvideo enc: choose VBV buffer size for mpeg4 when none is specified instead of failing
  mpegvideo enc: choose VBV buffer size for mpeg1/2 when none is specified instead of failing
  ffv1: fix array data types
  movenc: fix regression with yuyv caused by c5f23d
  libvorbisenc: fix afq delay setting
  aacenc: fix out of array writes
  ...

git-svn-id: http://svn.ad-group.adh/svn/native_client/native/ffmpeg/trunk@8942 4601b0d2-a6ba-4ff0-a3bb-59d57de6f651

+++ libavcodec/ffv1.c
@@ -1242,7 +1242,7 @@ static int encode_slice(AVCodecContext *c, void *arg){
     }
     if(!fs->ac){
         if(f->version > 2)
-            put_rac(&fs->c, (int[]){129}, 0);
+            put_rac(&fs->c, (uint8_t[]){129}, 0);
         fs->ac_byte_count = f->version > 2 || (!x&&!y) ? ff_rac_terminate(&fs->c) : 0;
         init_put_bits(&fs->pb, fs->c.bytestream_start + fs->ac_byte_count, fs->c.bytestream_end - fs->c.bytestream_start - fs->ac_byte_count);
     }
@@ -1313,9 +1313,8 @@ static int encode_frame(AVCodecContext *avctx, AVPacket *pkt,
 
     for(i=1; i<f->slice_count; i++){
         FFV1Context *fs= f->slice_context[i];
-        uint8_t *start = pkt->data + (pkt->size-used_count)*i/f->slice_count;
+        uint8_t *start = pkt->data + (pkt->size-used_count)*(int64_t)i/f->slice_count;
         int len = pkt->size/f->slice_count;
-
         ff_init_range_encoder(&fs->c, start, len);
     }
     avctx->execute(avctx, encode_slice, &f->slice_context[0], NULL, f->slice_count, sizeof(void*));
@@ -1678,7 +1677,7 @@ static int decode_slice(AVCodecContext *c, void *arg){
 
     if(!fs->ac){
         if (f->version == 3 && f->minor_version > 1 || f->version > 3)
-            get_rac(&fs->c, (int[]){129});
+            get_rac(&fs->c, (uint8_t[]){129});
         fs->ac_byte_count = f->version > 2 || (!x&&!y) ? fs->c.bytestream - fs->c.bytestream_start - 1 : 0;
         init_get_bits(&fs->gb,
                       fs->c.bytestream_start + fs->ac_byte_count,
@@ -1707,7 +1706,7 @@ static int decode_slice(AVCodecContext *c, void *arg){
     }
     if(fs->ac && f->version > 2) {
         int v;
-        get_rac(&fs->c, (int[]){129});
+        get_rac(&fs->c, (uint8_t[]){129});
         v = fs->c.bytestream_end - fs->c.bytestream - 2 - 5*f->ec;
         if(v) {
             av_log(f->avctx, AV_LOG_ERROR, "bytestream end mismatching by %d\n", v);
